requirements.txt
Django==5.2.6
whitenoise==6.11.0
gunicorn==20.1.0
python-dotenv==1.0.0
Pillow==11.3.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9
psycopg[binary]==3.2.3
